8-2-02

+++ src/main/java/study/study/model/entity/Item.java
@@ -9,6 +9,7 @@ import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -32,7 +33,7 @@ public class Item {
 
     private String title;
     private String content;
-    private Integer price;
+    private BigDecimal price;
     private String brandName;
 
     private LocalDateTime registeredAt;
